GCI-1072: Not all order confirmations have a delivery method.
