+++ src/main/java/uk/gov/companieshouse/itemhandler/email/CertifiedCopyOrderConfirmation.java
@@ -5,6 +5,7 @@ import java.util.List;
 public class CertifiedCopyOrderConfirmation extends OrderConfirmation {
     private List<CertifiedDocument> certifiedDocuments;
     private String totalFee;
+    private String deliveryMethod;
 
     public List<CertifiedDocument> getCertifiedDocuments() {
         return certifiedDocuments;
@@ -20,5 +21,13 @@ public class CertifiedCopyOrderConfirmation extends OrderConfirmation {
 
     public void setTotalFee(String totalFee) {
         this.totalFee = totalFee;
+    }
+
+    public String getDeliveryMethod() {
+        return deliveryMethod;
+    }
+
+    public void setDeliveryMethod(String deliveryMethod) {
+        this.deliveryMethod = deliveryMethod;
     }
 }
